+++ packages/tracing/config/config.ts
@@ -1,8 +1,21 @@
-export interface TracingConfig {
+export type TracingConfig = {
   serviceName: string;
   url: string;
-}
+};
+
+// type LightsepConfig = {
+//   collector: "lightstep";
+//   serviceName: string;
+//   url: string;
+//   headers: {
+//     "Lightstep-Access-Token": string;
+//     [key: string]: string;
+//   };
+// };
 
 export const defaultConfig = (serviceName: string): TracingConfig => {
-  return { serviceName, url: "http://localhost:14268/api/traces" };
+  return {
+    serviceName,
+    url: "http://localhost:9411/api/v2/spans",
+  };
 };